[REFACTOR] 기사 원문과 요약을 별도의 패키지로 분리
