+++ src/main/java/com/example/whiplash/delivery/email/SmtpEmailSender.java
@@ -1,8 +1,8 @@
 package com.example.whiplash.delivery.email;
 
 
-import com.example.whiplash.article.domain.document.SummarizedArticle;
-import com.example.whiplash.article.repository.SummarizedArticleRepository;
+import com.example.whiplash.article.summary.domain.document.SummarizedArticle;
+import com.example.whiplash.article.summary.repository.SummarizedArticleRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.mail.SimpleMailMessage;
