+++ src/main/java/com/example/whiplash/global/service/RedisStreamsTaskProducer.java
@@ -1,6 +1,6 @@
 package com.example.whiplash.global.service;
 
-import com.example.whiplash.article.service.ArticleTaskProducer;
+import com.example.whiplash.article.summary.service.ArticleTaskProducer;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
